src/customer-order-assistant/requirements.txt
# Customer Order Assistant MCP Server — Azure Functions
azure-functions>=1.21.0
azure-cosmos>=4.9.0
azure-identity>=1.17.0
aiohttp>=3.9.0
mcp[cli]>=1.26.0
pydantic>=2.0
python-dotenv>=1.0.0
